Replaced score type int with float

+++ code/Mimax/Minimax/IMinimaxResolver.h
@@ -11,7 +11,7 @@ class IMinimaxResolver
 public:
     virtual ~IMinimaxResolver() = default;
 
-    virtual int EvaluateState(TState const& state) = 0;
+    virtual float EvaluateState(TState const& state) = 0;
     virtual std::vector<TTransition> CollectTransitions(TState const& state) = 0;
     virtual TState TakeTransition(TState const& state, TTransition const transition) = 0; 
 };
